Últimas correções para incluir um novo concurso.

+++ src/main/java/br/com/guarasoft/conteudoprogramatico/orgao/persistence/Orgao.java
@@ -39,7 +39,7 @@ public class Orgao extends AbstractDao<Orgao, Long> implements OrgaoRepository {
 
 	@Override
 	public Orgao find() {
-		return super.find(id);
+		return super.find(id.longValue());
 	}
 
 }
